refactor: update types/xml match schema hierarchy
I had flattened this in my initial implementation, but it is just causing
me confusion as I update to the new schema and expand it. I expect
there will be a few smaller passes to get types fully aligned to the XML.

I also plan to match the type names exactly. I've dropped a 'Type'
suffix in a lot of cases that is also causing me some confusion.