fix: POST /account no longer supported by ESPM
With the release of the Test UI, ESPM has removed the API for
create accounts. This needs to be done through the UI now.
This commit removes the client's support for this feature and
modifies the tests to run without the assumption of a new
account being created on every test run.